thuyetminh_scr/requirements_streamlit.txt
streamlit>=1.28.0
yt-dlp>=2023.10.13
openai-whisper>=20231117
deep-translator>=1.11.4
google-generativeai>=0.3.0
pinecone>=2.2.4
pydub>=0.25.1
moviepy>=1.0.3
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pytube>=15.0.0
gtts>=2.3.2
pyttsx3>=2.90
edge-tts>=6.1.9
tqdm>=4.65.0
pathlib2>=2.3.7
plotly>=5.15.0 